requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2023.11.17
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.0
fastapi==0.109.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
openai==1.10.0
packaging==23.2
pluggy==1.4.0
pydantic==2.5.3
pydantic_core==2.14.6
pytest==7.4.4
pytest-asyncio==0.23.3
python-dotenv==1.0.1
sniffio==1.3.0
starlette==0.35.1
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.9.0
uvicorn==0.27.0
